Update README.md
https://github.com/dustinkrejci/Oceania-command

Oceania command is a all in one terminal. This does everything w/out all the tools of the trade and what can be found in the market. Forgoing premade to custom and nonchalant omi-matic: "all of one in ones' self". Forgoing the choaticiness of the many, to the singleness and choas of ones self. build, to build: what to build today in the empire of the Ocean?



All these tools work, and so does junk in the hands of a skill workman. all that junk is a rath of tools that others are floating on, what are floating in That Ocean? 

Ocean command is a tool in which it starts from the terminal in the roadmap and branches out to other tooling.

Oceania command is an single and isolated tool with its tool chain. but it has other tools and toolchains to operate  everything, including global entity organizations, done to a server on the darknet to run a minecraft server for 1million users, to cooperate business software and everything the internet could want and demand of Ocean command.

but if one whats to command a fleet as big as an armada of jet ski or even warships of compute with multi cloud global power behide it, you asked, then can do so, and it starts here in Oceania command. The tide is your to time, but its Oceania time to build and show you how, here is your workshop you never asked, but now soon to have like any masters builder workshop to build up the tools of a lifetime, but we like to shorted that timescale. 



Oceania command is also integrated into the web browser, called Oceania or alternatively Oceania Web browser.





Oceania command is integrated into  a custom kernel.

Oceania command is integrated into a server version of the o called, "Oceania command vessel", furthermore Oceania command has  the integrated custom kernels as mentioned briefly. 

Oceania command is integrated into the custom OS called Oceania OS with the mentioned kernel and a possible second OS.